# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FreightUtils API",
    "description": "Free freight calculation and reference data APIs. No authentication required. JSON responses. CORS enabled for all origins.",
    "version": "1.0.0",
    "contact": {
      "name": "FreightUtils",
      "url": "https://www.freightutils.com",
      "email": "contact@freightutils.com"
    },
    "license": {
      "name": "Free to use",
      "url": "https://www.freightutils.com/api-docs"
    }
  },
  "servers": [
    {
      "url": "https://www.freightutils.com",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/unlocode": {
      "get": {
        "summary": "UN/LOCODE transport location lookup",
        "description": "Search 116,000+ UN/LOCODE entries by name, code, country, or function type.",
        "operationId": "searchUnlocode",
        "tags": [
          "Reference"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search by name or code",
            "schema": {
              "type": "string"
            },
            "example": "rotterdam"
          },
          {
            "name": "code",
            "in": "query",
            "description": "Exact UN/LOCODE",
            "schema": {
              "type": "string"
            },
            "example": "NLRTM"
          },
          {
            "name": "country",
            "in": "query",
            "description": "ISO 2-letter country",
            "schema": {
              "type": "string"
            },
            "example": "GB"
          },
          {
            "name": "function",
            "in": "query",
            "description": "Function type",
            "schema": {
              "type": "string",
              "enum": [
                "port",
                "airport",
                "rail",
                "road",
                "icd",
                "border"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max results",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results"
          },
          "404": {
            "description": "Code not found"
          }
        }
      }
    },
    "/api/uld": {
      "get": {
        "operationId": "getULDs",
        "summary": "Air Freight ULD Types",
        "description": "Look up air freight ULD specifications. 15 types including AKE (LD3), PMC, temperature-controlled containers. Returns dimensions, weights, volume, aircraft compatibility.",
        "tags": ["Reference Data"],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "ULD code or slug (e.g. AKE, PMC). Omit to list all.",
            "required": false,
            "schema": { "type": "string" },
            "example": "AKE"
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category: container, pallet, or special.",
            "required": false,
            "schema": { "type": "string", "enum": ["container", "pallet", "special"] }
          },
          {
            "name": "deck",
            "in": "query",
            "description": "Filter by deck position: lower or main.",
            "required": false,
            "schema": { "type": "string", "enum": ["lower", "main"] }
          }
        ],
        "responses": {
          "200": { "description": "ULD specification(s)." },
          "404": { "description": "ULD type not found." }
        }
      }
    },
    "/api/vehicles": {
      "get": {
        "operationId": "getVehicles",
        "summary": "Vehicle & Trailer Types",
        "description": "Look up road freight vehicle and trailer specifications. 17 types covering articulated trailers, rigid trucks, and vans. Returns dimensions, payload limits, pallet capacity.",
        "tags": ["Reference Data"],
        "parameters": [
          {
            "name": "slug",
            "in": "query",
            "description": "Vehicle slug (e.g. standard-curtainsider). Omit to list all.",
            "required": false,
            "schema": { "type": "string" },
            "example": "standard-curtainsider"
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category: articulated, rigid, or van.",
            "required": false,
            "schema": { "type": "string", "enum": ["articulated", "rigid", "van"] }
          },
          {
            "name": "region",
            "in": "query",
            "description": "Filter by region: EU or US.",
            "required": false,
            "schema": { "type": "string", "enum": ["EU", "US"] }
          }
        ],
        "responses": {
          "200": { "description": "Vehicle specification(s)." },
          "404": { "description": "Vehicle type not found." }
        }
      }
    },
...
  },
...
}